chore: rename recovery to resilience across score module
