+++ backend/app/schemas/responses/activity/resilience.py
